+++ changelog.md
@@ -2,6 +2,9 @@ Codion Change Log
 =================
 
 ## 0.18.32
+### is.codion.common.model
+- UserPreferences.getUserPreference(key, defaultValue) now requires a non-null default value, can no longer return null.
+- MultiSelection.Indexes and Items override get() annotated with @NonNull.
 ### is.codion.swing.common.ui
 - FilterTable bug fixed, doubleClickAction no longer triggered if disabled.
 - ComponentBuilder.toolTipText(Observable<String>) added for a dynamic tool tip.